+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/deploy-tooling
 
+## 0.16.60
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/inquirer-common@0.8.0
+
 ## 0.16.59
 
 ### Patch Changes