+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.29
+
+### Patch Changes
+
+-   Updated dependencies [4053369]
+    -   @sap-ux/odata-service-inquirer@2.9.0
+    -   @sap-ux/inquirer-common@0.8.0
+    -   @sap-ux/abap-deploy-config-inquirer@1.6.30
+
 ## 0.2.28
 
 ### Patch Changes